pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rageval"
version = "0.1.0"
description = "Production-grade RAG evaluation framework with retrieval and generation metrics"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Manesh Thiagarajan" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pydantic>=2.0",
    "structlog>=23.0",
    "duckdb>=0.9",
    "llama-index-core>=0.11",
    "llama-index-llms-openai>=0.2",
    "llama-index-embeddings-openai>=0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.9",
]

[tool.hatch.build.targets.wheel]
packages = ["src/rageval"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--strict-markers"

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
disallow_untyped_defs = true
packages = ["rageval"]
mypy_path = "src"

[tool.coverage.run]
source = ["src/rageval"]

[tool.coverage.report]
fail_under = 75
